README + Deployment + SDK Updates

+++ start-idsvr.sh
@@ -1,29 +1,15 @@
 #!/bin/bash
 
 ####################################################################################################
-# Run the Curity Identity Server in Docker on the local computer, preconfigured for the code example
-# Please ensure that the following resources are installed before running this script:
-# - Docker Desktop
-# - The envsubst tool (`brew install gettext` on MacOS)
-#
-# If you want to expose the local instance of the Curity Identity Server via ngrok, then the following
-# need to also be installed:
-# - ngrok
-# - The jq tool (`brew install jq` on MacOS)
+# A developer deployment to run the Curity Identity Server in Docker on the local computer.
+# The deployment provides a working configuration and includes native passkeys automation.
+# - https://curity.io/resources/learn/mobile-logins-using-native-passkeys/
 ####################################################################################################
 
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
 #
-# By default the Curity Identity Server will use the Android emulator's default host IP.
-# Set USE_NGROK to true and a dynamic NGROK base URL will be used automatically.
-#
-USE_NGROK='false'
-BASE_URL='https://10.0.2.2:8443'
-EXAMPLE_NAME='haapi'
-
-#
-# First check prerequisites
+# First check for a license
 #
 if [ ! -f './license.json' ]; then
   echo 'Please copy a license.json file into the root folder of the code example'
@@ -31,37 +17,72 @@ if [ ! -f './license.json' ]; then
 fi
 
 #
+# Then validate or default parameters
+#
+EXAMPLE_NAME='haapi'
+if [ "$USE_NGROK" != 'true' ]; then
+  USE_NGROK='false'
+  if [ "$IDSVR_HOST_NAME" == '' ]; then
+    echo 'You must supply an IDSVR_HOST_NAME for the Curity Identity Server'
+    exit 1
+  fi
+fi
+
+#
 # Download mobile deployment resources
 #
-rm -rf deployment
-git clone https://github.com/curityio/mobile-deployments deployment
-if [ $? -ne 0 ]; then
-  echo 'Problem encountered downloading deployment resources'
-  exit 1
+if [ ! -d ./deployment ]; then
+  rm -rf deployment
+  git clone https://github.com/curityio/mobile-deployments deployment
+  if [ $? -ne 0 ]; then
+    echo 'Problem encountered downloading deployment resources'
+    exit 1
+  fi
 fi
+
+#
+# Override the HAAPI default configuration settings
+#
+cp config/docker-template.xml deployment/haapi/example-config-template.xml
 
 #
 # Run an automated deployment of the Curity Identity Server
 #
 cp ./license.json deployment/resources/license.json
-./deployment/start.sh "$USE_NGROK" "$BASE_URL" "$EXAMPLE_NAME"
+if [ "$USE_NGROK" == 'true' ]; then
+  ./deployment/start.sh 'true' '-' "$EXAMPLE_NAME"
+else
+  ./deployment/start.sh 'false' "https://$IDSVR_HOST_NAME:8443" "$EXAMPLE_NAME"
+fi
 if [ $? -ne 0 ]; then
   echo 'Problem encountered deploying the Curity Identity Server'
   exit 1
 fi
 
 #
-# Inform the user of the Curity Identity Server URL, to be copied to configuration
+# Get the final Curity Identity Server URL
 #
-RUNTIME_BASE_URL="$(cat ./deployment/output.txt)"
-echo "Curity Identity Server is running at $RUNTIME_BASE_URL"
+export IDSVR_BASE_URL="$(cat ./deployment/output.txt)"
+echo "Curity Identity Server is running at $IDSVR_BASE_URL"
 
 #
-# If using ngrok, configure more advanced settings
+# Configure the Android app to use the Curity Identity Server's base URL
 #
-if [ "$USE_NGROK" == 'true' ]; then
-  ./configure-app.sh
-  if [ $? -ne 0 ]; then
-    exit 1
+function replaceTextInFile() {
+
+  FROM="$1"
+  TO="$2"
+  FILE="$3"
+  
+  if [ "$(uname -s)" == 'Darwin' ]; then
+    sed -i '' "s/$FROM/$TO/g" "$FILE"
+  else
+    sed -i -e "s/$FROM/$TO/g" "$FILE"
   fi
+}
+
+IDSVR_HOST_NAME="${IDSVR_BASE_URL#https://}"
+replaceTextInFile '10.0.2.2:8443' "$IDSVR_HOST_NAME" './app/src/main/java/io/curity/haapidemo/Configuration.kt'
+if [ "$USE_NGROK" == 'true' ]; then
+  replaceTextInFile '10.0.2.2:8443' "$IDSVR_HOST_NAME" './app/src/main/res/values/strings.xml'
 fi
